*** empty log message ***

+++ web_module/trunk/src/main/com/topcoder/web/tc/controller/request/development/TCCC05ProjectDetails.java
@@ -96,7 +96,7 @@ public class TCCC05ProjectDetails extends StatBase {
                 if(a1.getScoreDouble() == a2.getScoreDouble()) {
                     return a1.getHandle().compareToIgnoreCase(a2.getHandle());
                 }
-                return a2.getScoreDouble() - a1.getScoreDouble();
+                return (int)(a2.getScoreDouble() - a1.getScoreDouble());
             }
             
             return a2.getPoints() - a1.getPoints();
